CODES Core: Separate network models from base
This addresses issue #122. Moves base implementations to their
own folder in src/networks/model-net/ to make clear distinction
between actual network models and the base implementations required
to run those network models.

This also moves a header file out of the src directory and into the
codes/ header directory where it belongs.
