+++ src/Makefile.subdir
@@ -90,6 +90,7 @@ nobase_include_HEADERS = \
 	codes/model-net-method.h \
 	codes/model-net-lp.h \
 	codes/model-net-sched.h \
+	codes/model-net-sched-impl.h \
 	codes/model-net-inspect.h \
 	codes/connection-manager.h	\
 	codes/net/common-net.h \
@@ -158,7 +159,7 @@ src_libcodes_la_SOURCES =  \
 	src/workload/methods/codes-iomock-wrkld.c \
 	codes/rc-stack.h \
 	src/util/rc-stack.c \
-	src/networks/model-net/model-net.c \
+	src/networks/model-net/core/model-net.c \
 	src/networks/model-net/common-net.c \
 	src/networks/model-net/simplenet-upd.c \
 	src/networks/model-net/torus.c \
@@ -171,10 +172,9 @@ src_libcodes_la_SOURCES =  \
 	src/networks/model-net/fattree.c \
 	src/networks/model-net/loggp.c \
 	src/networks/model-net/simplep2p.c \
-	src/networks/model-net/model-net-lp.c \
-	src/networks/model-net/model-net-sched.c \
-	src/networks/model-net/model-net-sched-impl.h \
-	src/networks/model-net/model-net-sched-impl.c
+	src/networks/model-net/core/model-net-lp.c \
+	src/networks/model-net/core/model-net-sched.c \
+	src/networks/model-net/core/model-net-sched-impl.c
 
 src_libcodes_mpi_replay_la_SOURCES = \
   src/network-workloads/model-net-mpi-replay.c
