+++ src/networks/model-net/core/model-net-lp.c
